View additional details of an event when selected
Moved the selected item logic to the main application Update func.
This is due to the lack of access to the application model from the
UpdateFunc within the list component.

In doing so, selecting an item changes the view, whereas the purpose of
the UpdateFunc for the list is to act upon a particular item within the
list, i.e. deleting an item, etc.

In the future, editing an item would most likely require logic within
the main Update func as well, since that would change to a different
view for the form.

+ Clean up unneeded list actions