feat: Add delete functionality
We put the delete within the list update func, as the list updateFunc
only has the scope of the event. We could also put it
within the main `Update()` and use the `model.selected`
attribute + MODE to accomplish the same.

`d` also conflicts with page down, so only SHIFT+d for key binding.